+++ NEWS.md
@@ -1,5 +1,14 @@
 # mlr3torch (development version)
 
+## Features
+
+* Feat: Added learners for the remaining image classification networks of `torchvision`:
+  ConvNeXt (`classif.convnext_*`), EfficientNet (`classif.efficientnet_b0` to
+  `classif.efficientnet_b7`), EfficientNetV2 (`classif.efficientnet_v2_{s,m,l}`),
+  Inception v3 (`classif.inception_v3`), MaxViT (`classif.maxvit`), MobileNetV3
+  (`classif.mobilenet_v3_{large,small}`), Vision Transformers (`classif.vit_*`) and
+  Wide ResNet (`classif.wide_resnet{50_2,101_2}`).
+
 ## Bug fixes
 
 * `logical()` features are now encoded as `c(1, 2)` by the